binance_bot_rsi/requirements.txt
#pip install -r requirements.txt
python-binance
numpy
requests
ta-lib
time
